Create submodule of `device::trace` for recording support
